Corrigiendo código 26/11

+++ api-peliculas/src/main/java/com/salesianos/dam/apipeliculas/service/DirectorService.java
@@ -17,6 +17,7 @@ import java.util.Locale;
 @Service
 public class DirectorService {
 
+    private static final Long director_desc = 0L;
     private final DirectorRespository directorRespository;
 
     public List<Director> getAll(){
@@ -28,17 +29,17 @@ public class DirectorService {
     }
 
     public Director create(DirectorRequestDTO dto){
-        if(!StringUtils.hasText(dto.nombre())) throw new IllegalArgumentException("Falta el campo del nombre del director");
 
-        if(LocalDate.now().getYear() - dto.anioNacimiento()<18) throw new DirectorMenorEdadException("El director no puede ser menor de edad");
+        DirectorRequestDTO.validarDTO(dto);
+        Director.comprobarEdad(dto.anioNacimiento());
 
         return directorRespository.save(dto.toEntity());
     }
 
     public Director edit(Long id,DirectorRequestDTO dto){
-        if(!StringUtils.hasText(dto.nombre())) throw new IllegalArgumentException("Falta el campo del nombre del director");
 
-        if(LocalDate.now().getYear() - dto.anioNacimiento()<18) throw new DirectorMenorEdadException("El director no puede ser menor de edad");
+        DirectorRequestDTO.validarDTO(dto);
+        Director.comprobarEdad(dto.anioNacimiento());
 
         return directorRespository.findById(id)
                 .map(d -> {
@@ -51,7 +52,13 @@ public class DirectorService {
     }
 
     public void delete(Long id){
+        if(id.equals(director_desc)) throw new IllegalArgumentException("No se puede eliminar");
+
         Director d = directorRespository.findById(id).orElseThrow(()-> new DirectorNoEncontradoException(id));
+        Director desc = directorRespository.findById(0L).orElseThrow(() -> new DirectorNoEncontradoException("No existes en la base de datos dicho director"));
+
+        d.eliminarDirector(desc);
+
         directorRespository.delete(d);
     }
 }
